24/07/2023 - 16:47

+++ website/views.py
@@ -398,18 +398,22 @@ def categoriaCadastrar(nome):
 def categoriaDeletar(id):
     categorias = Categoria.query.all()
     tamanhos = Tamanho.query.all()
-
+    exixte_em_produtos = Produto.query.filter_by(categoria=id).all()
     get_categorias = Categoria.query.get(id)
-    
-    if request.method == 'GET':
-        if get_categorias:
-            db.session.delete(get_categorias)
-            db.session.commit()
-            return render_template("categorias_listar.html", 
-                user=current_user,
-                data=data, 
-                categorias=categorias,
-                tamanhos=tamanhos)
+
+    if exixte_em_produtos:
+        flash('NÃO PODE SER EXCLUÍDO!!!')
+    else:
+        if request.method == 'GET':
+            if get_categorias:
+                db.session.delete(get_categorias)
+                db.session.commit()
+                return redirect(url_for('views.categorias'))
+                # return render_template("categorias_listar.html", 
+                #     user=current_user,
+                #     data=data, 
+                #     categorias=categorias,
+                #     tamanhos=tamanhos)
         
     return render_template("categorias_listar.html", 
         user=current_user,
